Revert "Optimize systemd dependency for containers"
It was generally a mistake to use `%systemd_ordering` in those
circumstances, as the tmpfiles stuff doesn't work and populate the
`/tmp` and `/run` directories. The reason why the packages install is
that all the commands in the macros have `|| :` at the end, so their
error state is ignored.

Something to re-consider for future attempts to fix/improve that:
If the tmpfiles and such were split out into a subpackage that'd be
required if and only if the kernel was installed or being installed,
that might work.
Related-Bug: #1804822
This reverts commit b3eddbe3326ded8269f199aa1bff008e2e094a83.

Change-Id: I841bc0d994c01c6ed16774faa9d9ebe8abcecca4