Update patch set 1
Patch Set 1:

Two new dependencies have been introduced in [1] which are not in CentOS or RDO, grpc and protobuf[1]. The good news is that they are probably are in Fedora, so it may be easy to import them in RDO (if we are lucky enough). Could you follow the process in [2] to add the dependencies in RDO Train?. Until ready, i'll propose to pin kuryr-kubernetes to the last known-good commit in master.

[1] https://opendev.org/openstack/kuryr-kubernetes/commit/48136ab034e313c6437e1c46161bc8078762965a
[2] https://www.rdoproject.org/documentation/requirements/#adding-a-new-requirement-to-rdo

Patch-set: 1
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0