Update patch set 1
Patch Set 1:

> > The build failed because of code-block directive missing.
 > > The solution is to replace ".. code-block::" by "::".
 > > Upstream don't test doc generation in py2 anymore because ussuri
 > is
 > > py3 only.
 > > We have two solutions:
 > > - make a patch upstream
 > > - deactivate doc generation for py2
 > 
 > If the upstream patch only requires adding a parameter to the
 > code-block in one place, I think we could go for it, as it will
 > make code cleaner. Otherwise, let's deactivate doc generation for
 > py2.

Issue is not specific to py3, but related to old version of sphinx that we have in RDO - 1.8.4. So we need to look at upgrading sphinx, to have a quick fix for now we can disable warning as error in doc build and in parallel look for upgrading sphinx as upstream u-c https://github.com/openstack/requirements/blob/master/upper-constraints.txt#L473-L477, also need to check compatibility of that version with py2.7.

Patch-set: 1